Textures and item creation
Added a few images and textures for a couple items and other player models
Added in a way to create items. press 1 to spawn a wood log at cursor

+++ core/src/com/rusd/tddwd/entity/components/DrawableComponent.java
@@ -16,8 +16,8 @@ public class DrawableComponent extends Component {
 	public Animation<TextureRegion> animation;
 	public float animTime;
 	
-	public int xOffset = -2;
-	public int yOffset = -2;
+	public float xOffset = -2;
+	public float yOffset = -2;
 	public String current;
 	public Map<String, TextureRegion> map = new HashMap<>();
